:truck: 修改推送 UUID 的接口为 checkUserTokenController
